Rework fileserver tag handling
There were a few nasty corner cases regarding flushes in the previous tag handling. Specifically, if a request was made, then flushed, and a new request made in its place at the right timing, the response for the first request would be written instead of the response to the second.

To fix this, a form of context object (requestState) has been introduced in order to store the cancel state, rather than rely on the validity of a tag number as response at a given time.